renaming player page name
